Refactor pattern length code into a package; add keymap to keygraphmap
