+++ lengthpattern/keymap.go
@@ -1,0 +1,63 @@
+package lengthpattern
+
+// Used to make KeyGraphMap.Get(r rune) use either normal symbol or shift symbol for arg r.
+type keyMap struct {
+	keyMap map[rune]key // normalSymbols index to map
+	shiftedKeyMap map[rune]key // shiftedSymbols as index to map
+}
+
+func NewKeyMap(matrix [][]*keyNode) *keyMap {
+	numRow       := "`1234567890-= "
+	numRowShift  := `~!@#$%^&*()_+ `
+	topRow       := ` qwertyuiop[]\`
+	topRowShift  := ` QWERTYUIOP{}|`
+	homeRow      := ` asdfghjkl;'  `
+	homeRowShift := ` ASDFGHJKL:"  `
+	botRow       := ` zxcvbnm,./   `
+	botRowShift  := ` ZXCVBNM<>?   `
+
+	km := &keyMap{
+		keyMap:        make(map[rune]key, len(numRow) + len(topRow) + len(homeRow) + len(botRow)),
+		shiftedKeyMap: make(map[rune]key, len(numRow) + len(topRow) + len(homeRow) + len(botRow)),
+	}
+	km.PushString(numRow, numRowShift)
+	km.PushString(topRow, topRowShift)
+	km.PushString(homeRow, homeRowShift)
+	km.PushString(botRow, botRowShift)
+	return km
+}
+
+// Pushes a key to the structure.
+func (km *keyMap) Push(k key) {
+	km.keyMap[k.Symbol] = k
+	km.shiftedKeyMap[k.ShiftSymbol] = k
+}
+
+// Pushes keys composed of symbols and shiftSymbols to the structure.
+func (km *keyMap) PushString(symbols, shiftSymbols string) {
+	for i := 0; i < len(symbols); i++ {
+		km.Push(key{rune(symbols[i]), rune(shiftSymbols[i])})
+	}
+}
+
+// Gets shift symbol if normal symbol passed, and vice versa.
+func (km keyMap) AntiSymbol(r rune) rune {
+	if anti, ok := km.keyMap[r]; ok {
+		return anti.ShiftSymbol
+	}
+	if anti, ok := km.shiftedKeyMap[r]; ok {
+		return anti.Symbol
+	}
+	return 0
+}
+
+// Gets key regardless if r is shift or non-shift
+func (km keyMap) Get(r rune) key {
+	if key, ok := km.keyMap[r]; ok {
+		return key
+	}
+	if key, ok := km.shiftedKeyMap[r]; ok {
+		return key
+	}
+	return key{}
+}
